refactor: rename `instructions` to `custom-instructions`
