+++ packages/website/src/router.tsx
@@ -2,10 +2,10 @@ import { Route, Switch } from 'wouter'
 
 import DocsLayout from './components/DocsLayout'
 import Header from './components/Header'
+import Instructions from './docs/features/custom-instructions/page'
 // Features pages
 import CustomTools from './docs/features/custom-tools/page'
 import DataMasking from './docs/features/data-masking/page'
-import Instructions from './docs/features/instructions/page'
 import ModelIntegration from './docs/features/model-integration/page'
 import SecurityPermissions from './docs/features/security-permissions/page'
 import BestPractices from './docs/integration/best-practices/page'
@@ -80,7 +80,7 @@ export default function Router() {
 				</div>
 			</Route>
 
-			<Route path="/docs/features/instructions">
+			<Route path="/docs/features/custom-instructions">
 				<div className="min-h-screen bg-white dark:bg-gray-900">
 					<Header />
 					<DocsLayout>
